Rename HelpScout/Beacon
One of the hardest problems in programming. Name changed to represent what the class is for now that we have removed the beacon. Hopefully.

[MAILPOET-5551]
