+++ mailpoet/lib/DI/ContainerConfigurator.php
@@ -351,8 +351,7 @@ class ContainerConfigurator implements IContainerConfigurator {
     $container->autowire(\MailPoet\Form\Templates\TemplateRepository::class);
     $container->autowire(\MailPoet\Form\Util\Styles::class);
     $container->autowire(\MailPoet\Form\Util\CustomFonts::class)->setPublic(true);
-    // Helpscout
-    $container->autowire(\MailPoet\Helpscout\Beacon::class)->setPublic(true);
+    $container->autowire(\MailPoet\SystemReport\SystemReportCollector::class)->setPublic(true);
     // Homepage
     $container->autowire(\MailPoet\Homepage\HomepageDataController::class)->setPublic(true);
     // Listing
